Crate fasten_kmer

Counts kmers. Each line is a kmer with two columns separated by tab: kmer, count

Examples

cat testdata/four_reads.fastq | fasten_kmer -k 15 > 15mers.tsv

Usage

Usage: fasten_kmer [-h] [-n INT] [-p] [-v] [-k INT]
Options:
   -h, --help          Print this help menu.
   -n, --numcpus INT   Number of CPUs (default: 1)
   -p, --paired-end    The input reads are interleaved paired-end
   -v, --verbose       Print more status messages
   -k, --kmer-length INT
                       The size of the kmer

Functions

count_kmers

Read fastq from stdin and count kmers

kmers_in_str

Read a str of nucleotides and count kmers. If should_revcomp is true, then will also count kmers on the opposite strand.

main

revcomp

reverse-complement a dna sequence

switch_base

Complementary nucleotide for ACTGUN, case insensitive